supplemental material/print-a-sketch_backend/requirements.txt
kivy
opencv2
numpy
serial